Delete old UI directory
Rather than running a separate initial vite command and then moving
things over, I think it'll be easier to just replay our changes, given
they're a pretty small set so far. The diffs would also be hard to
follow if I deleted the result of create-react-app and added the result
of vite in one go.