Adds an assertion that mock fetch was not called
I don't like doing this in a teardown function, but I think we lack good
alternatives given react-query's handling of any errors thrown while
attempting to fetch. I left the mock behavior in place, because it could
catch other usage of fetch, and the message is still present in the
output and clear. I couldn't present any more descriptive an error via
Jest without janky workarounds or adding a new package.

App.test.tsx fails at this commit to prove that the test works. Removing
in next commit, and will squash and merge.